pyproject.toml
[project]
name = "milkbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame>=2.6.1",
    "pyserial>=3.5",
]
